reading, photos: use YAML for index, instead of markdown tables
Markdown tables are hard to manage and scale horribly. Instead use YAML
to define the index for photos and reading, for now. It simply reads
YAML and then forwards that to Go templating. The templates in
danishpraka.sh/ then render the different YAML fields for an item (book
or photos).

Also adds a markdown file.

Signed-off-by: Danish Prakash <contact@danishpraka.sh>